//Actor類: //1.角色可以朝目標移動(彈道定點會抖動，須修正)-->moveToFlag方法 //2.新增方向屬性和標示是否為敵軍的屬性 //3.建構子修改成需要傳入第三個參數-是否為敵軍。 //4.新增 changeDir方法
//Rect類:
//1.將int改成flot，並增加offSet中心點座標的方法
//Tank1及Enemy1角色切圖
//建構子屬性初始化

//Flag
//int的地方改成float

+++ src/utils/Flag.java
@@ -30,7 +30,7 @@ public class Flag implements GameKernel.PaintInterface,GameKernel.UpdateInterfac
 
     @Override
     public void paint(Graphics g) {
-        g.drawImage(image,painter.centerX(),painter.centerY(),null);
+        g.drawImage(image,(int)painter.centerX(),(int)painter.centerY(),null);
     }
 
     @Override
